backend/requirements.txt
Flask
Flask-CORS
Flask-Session
python-dotenv
werkzeug
cs50
